Submit direct zone updates to our API by AJAX
This prevents updates from being lost if the submission fails for any
reason. Since we gave up any pretense at being a standard HTML form
long ago, we are not losing anything by doing this, and we gain in not
losing all of a user's changes if one of their changes has an error in
it.

If we receive an error, display it to the user just above the submit
button.

Resolves: #9