Perform DELETE actions before REPLACE actions
When changing the name or type of an RRSet, we perform a REPLACE action for
the new RRSet and a DELETE action for the old one. We need to make it do
DELETE before REPLACE, otherwise we will hit the CNAME singleton check in
PowerDNS 4.1 when trying to change eg. an A record to a CNAME.

Resolves: #34