Adding alternative user data source based on PHP_AUTH_ variables.
As an alternative to LDAP, user data can now be found in the following PHP variables:
PHP_AUTH_NAME - the full name of the user
PHP_AUTH_EMAIL - the email address of the user
PHP_AUTH_GROUPS - the list of groups the user is a member of, separated by spaces

A configuration section [php_auth] was added and a `enabled` flag was added to [ldap] section.
Only one data source can be enabled, either ldap or php_auth.
The `admin_group` option defines the criteria for selection of administrative users.
Only the users that have the `admin_group` value in their group list will be created on the database as administrators.

The PHP variables mentioned have to be passed from the web server to the PHP application, like PHP_AUTH_USER.
The single sign on module on the web server has to be configured to get the required data and pass the values to such variables.