Stop using badges for things that don't need attention
The <span class="badge"> in Bootstraps is intended for things such as
"highlight[ing] new or unread items". It isn't really intended for showing the
total number of rows. Since there is no more appropriate class, I'm removing
the styling altogether.

This is a precursor to another commit soon to follow that will add some badges
for things that genuinely need attention.

See also: #110