Apply data-* attributes to table columns for user defined styles
This allows an administrator to set styles on these columns, for example
to highlight different Classifications. Class names are sanitized to
only contain letters and numbers and are prefixed by their column type.
The rules are to be placed in public_html/site.css, which does not exist
by default.

e.g. to highlight columns where Classification is "internal":

  tr[data-account="internal"]>td.account {
    background-color: #d9edf7;
  }
  tr[data-account="internal"]:hover>td.account,
  tr[data-account="internal"]>td.account:hover {
    background-color: #c4e3f3;
  }