Don't raise and exception when trying to parse date without microseconds
When SELECTing a date from the postgres database, and that date's
microsecond part is exactly zero, it will be omitted from the string
representation. This causes DateTime::createFromFormat to return false
instead of the date.

To reproduce this bug, create and then edit a zone (so a changeset entry
exists), then set its change_date like so:

    update changeset set change_date = '2021-11-23 10:37:38.000000' where id = 1;